Created devcontainer
The devcontainer requires the VSCode devcontainer extension and docker
to be installed. To utalize the devcontainer, open the repo in VSCode
with the extension installed and run the "Dev Containers: Reopen in Container"
command.

The Devcontainer is an ubuntu system with several utlities installed,
including: git, zsh, curl, and more and also comes with a few VSCode
enxtensions: prettier and eslint, which I imagine will be useful for
this project.

The Prettier extension is a javascript autoformatter that can be configured
using a ".prettierrc.json" file, and eslint can be configured using a
".eslintrc.json" file. Neither of these have been created in this commit.